requirements.txt
crewai
langchain
pillow
streamlit
graphviz
serper-dev